+++ src/Braintree/Properties/AssemblyInfo.cs
@@ -32,5 +32,5 @@ using System.Runtime.InteropServices;
 //      Build Number
 //      Revision
 //
-[assembly: AssemblyVersion("5.16.0.0")]
-[assembly: AssemblyFileVersion("5.16.0.0")]
+[assembly: AssemblyVersion("5.17.0.0")]
+[assembly: AssemblyFileVersion("5.17.0.0")]